Add files via upload
Uploading a file with my own macro doing all the stocks analysis for 2018. My macro calculates Total Volume and the Return for each ticker (company) using just one loop through all the rows in the worksheet 2018 and puts the received data to the output worksheet All Stocks Analysis. I will use arrays and nested loops in a later code as suggested but I think it's inefficient to use in this case.