Add files via upload
Uploading a file with a macro described in chapter 2.3.3 doing all the stocks analysis for 2018. 
This  macro calculates Total Volume and the Return for each ticker (company) using array to store tickers and nested loop to loop through all the rows in the worksheet (inner loop) for every ticker from a predefined array (outer loop).